function で関数一覧を表示する機能を追加
計算履歴機能で範囲外アクセスで落ちる問題を修正

mem[index]からmem(index)に変更
(それに伴い一部のトークンを削除)

定義されていない変数xでmem(x)とした場合エラーで落ちる問題を修正

+++ src/memory.h
@@ -1,4 +1,4 @@
 #pragma once
 
 void pushMemory(double value);
-double getMem(unsigned int index);
+double getMem(double index);
